:ambulance: Replace HEAD calls with GET for OIDC auth endpoint
The OpenID Connect standard prescribes nothing about having to support
HEAD requests, so we can't be certain that all OIDC providers support
this, nor can we make a reasonable demand to support this at the
infrastructure level.

This replaces the HEAD requests with a GET, which should be guaranteed
to work.

Backport-of: open-formulieren/open-forms#3194