:recycle: [#3193] -- Clean up JCC plugin exception handling
This allows to focus on the actual SOAP calls and DRY's the logging
(until we have structlog) + default value fallback on errors.

The mechanism is rather simple - errors are surpressed and logged by
the context manager, which classifies exceptions in two types:

* fall back to default value (because of errors we can't do anything
  about)
* crash because of programming mistake on our end

The context manager raises one of both error types, which are picked
up by the decorator to decide if a default value is returned or the
error is left to bubble up.