:sparkles: [#3193] -- implement qmatic get required customer fields
Qmatic does not have an endpoint for this, only the guideline that you
should at least include one field. This is expressed in the model
field blank=False option.

The available list of fields is defined as an enum, and can be configured
in the qmatic admin page to select the appropriate fields. The default
includes typical fields used in NL to identify people.