:fire: Remove CI job for MacOS deps installation
Workarounds exists, but these are the responsibility of the people
who really want to use MacOS as development environment. We do not
officially support deploying on or developing with MacOS.

The failing dev installation is related to the python3-saml stack,
which requires xmlsec, which requires libxml < 1.3. The default brew
recipes install libxml 1.3.0+ and the old formula's are no longer
available without some annoying manual workarounds.