♻️[#3179] Refactor registration plugin configuration
Because form logic has can change which backend-config combination will
be executed, `registration_backend` is now a property on the submission,
not the form.

DRF could use a bit "Parse don't validate": at `validate` time
(especially during a PATCH operation) not everything is known. Like
which plugin, and what the type of the options for that plugin are.
This is the shotgun validation that results: bits of validation and
re-validation all over the place.