:sparkles: [#2471] -- implement create appointment endpoint
The Create appointment endpoint is called in the SDK when
the end-user confirms the data on the summary screen. It
performs a final input-data validation, against the
selected appointment plugin.

When all is well, the appointment data is stored in the DB,
together with the plugin that was used so if configuraton
is changed, then the data can be looked up again in the
correct plugin.

Validation errors are returned to the frontend which can
display them again, if applicable.