:pushpin: Unpin Python from 3.10.9 to 3.10
This was originally pinned due to a threading shutdown bug in CPython,
see #2748. Since then Python 3.10.11 and 3.10.12 have been released,
containing security fixes.

Not specifying the patch version in docker images and CI pipelines
grabs the latest patch version for the minor version, which gives us
the latest bug fixes 'for free'. This is preferrable over the risk
of new bugs breaking our application like in #2748.