:bug: [#3309] Add missing appointment configuration migration
The appointment plugins were normalized to be proper plugins using the
decorator approach, which caused their configuration paths/identifiers
to change. However, there was no data migration to map from the old
configuration paths to the new identifiers.

This fix is *not* backported because it deals with migrations. There
is a simple workaround available for DevOps to update the value in
the database to either qmatic or jcc, which were the supported plugins
before.