:label: Make registry/plugin definitions generics
These help in returning the appropriate plugin class for further usage.

Note that the string version of type annotations is required to resolve
circular imports. This in turn leads to Flake8 reporting F401 for
unused imports, but that's because it cannot solve the generics
and typevars, see https://github.com/PyCQA/pyflakes/issues/571