:sparkles: [#3333] Allow making additional fields required in JCC configuration
Clients sometimes need to make fields optional in JCC when staff are
creating appointments on behalf of customers, where they don't know
those details or they're not relevant. However, when an online
appointment is created by the customer, these fields should be
mandatory input to guarantee the best 'customer journey'.

This allows marking additional fields are required that are
optional in JCC. It does not allow marking required fields in JCC
optional (as that would crash when creating the appointment).