✨[#1889] Add current_year static variable
Closes #1889

Having current_year available as a variable solves the problem of
Haarlemmermeer of not being able to "block"/validate based on an integer
relative to the current year. They can adapt their current forms to use
this, so they won't need to change the forms on each January 1st
anymore.

I think the wish for a picker is a distraction from this core problem.
But should one really feel the need for a year picker, with this
`current_year` one can construct a one with a Select component with
variable options with an option expression like this:

```json
{
  "merge": [
    { "-": [{ "var": "current_year" }, 3] },
    { "-": [{ "var": "current_year" }, 2] },
    { "-": [{ "var": "current_year" }, 1] },
    { "var": "current_year" },
    { "+": [{ "var": "current_year" }, 1] },
    { "+": [{ "var": "current_year" }, 2] },
    { "+": [{ "var": "current_year" }, 3] }
  ]
}
```

This solution evades all kinds of combinations FormIO fun and
JavaScript's ill-conceived Date API.

PS: use `{{ current_year|stringformat: "d" }}` to diplay the value in templates
without thousands separator.