:recycle: [#3406] Lift translate-function creation up into FormioNode
The translate function is required at the component level to fix
the labels because the value formatters do not have access to
translations. Therefore, the highest node we can lift this up to
is the FormioNode, as that is called by the SubmissionStep node.

This node still needs access to the step, because it must be
able to look up the formio configuration from the related
form step and loop over each component. The translate function
only needs to be determine once and can then be passed around.

Its presence indicates that translating is necessary, as a
None value indicates translating can be skipped.