:art: Rewrite stuf/soap service migration history
Turn the separate-database-and-state operations into real ones,
sorting out the correct order of migrations in the
squashed migrations.

Existing installations are already in the correct state, and
the squashed migrations are executed for fresh installs which
sets up things the right way from the get-go rather than
using advanced migration techniques.