:recycle: [#3489] Implement a better API to pass client init kwargs
The factory function to instantiate a particular client class usually
knows which additional kwargs subclasses/specific clients need, so
it makes more sense to allow kwargs to be passed to the classmethod
to construct a client instance.

This makes the factory/manual approach a bit more equivalent and less
magical or annoying where certain kwargs are just passed through.