:recycle: [#3489] Refactor the BAG client implementation
Refactored the client to be based on our api_client.APIClient base,
instead of using the zds-client approach which is YAML/OAS based.

This means that you can now use/configure the BAG service without
needing the OpenAPI specification.

Client and config check have been separated, in line with the kadaster
config check.

The tests have been cleaned up - there was unnecessary indirection for
mocking and configuration management. The get_solo calls are now
properly mocked so that they don't hit the cache or DB (which is also
nice for test performance).

Finally, the OPENFORMS_LOCATION_CLIENT is removed. This was
undocumented (and also a poor interpretation of some what-if that did
not materialize) and considered private API. This removes some more
indirection and improves the searchability of the code base.