:recycle: [#3489] Refactor kadaster location server client
The pre-requests ZGW client stuff did not seem to actually be used
anywhere, so we're not losing functionality (the context is never
being set).

After some discussion with Silvia about how this is used (solely in a
request auth context), we'll probably add first-class support to
APIClient for it so we can build the functionality in in a generic
way. This could *potentially* be solved through the factories, as
relying on the auth kwarg seems to be the way forward.

Note also that the request error suppression can probably be solved in
a more elegant way, but let's tackle that at the end of this refactor.