:card_file_box: [#3489] Set up migration for KVK service normalization
This takes the common base of the zoeken/profile API services
and grabs or creates a new service for the common api root,
which is configured on the new (and later only) field for
the KVK configuration.

The API client will then explicitly build up the API URLs
to consume instead of relying on OpenAPI specs.