:recycle: [#3489] Refactor KVK basisprofiel client
There will be a single KVK client definition as they all share the same
API root. The KVK prefill and client tests for basisprofiel usage are
updated with the new client, based around the api_client.APIClient
structure.

Tests are cleaned up and prepared to be replaced with VCR tests even.

Next up is replacing the search API client.