:art: [#3481] Simplify tests more
The full integration test with posting to the objects API
doesn't need assertions that aren't related to the escaped
data -> we can remove the result checking and set up the
request mocks to always automatically reflect the input
record data to avoid this going out of sync when the
template/posted data changes.

The second aspect puts the assertion inside the self.subTest
block, as test failures need to be recorded inside that
context for proper reporting back.