:recycle: [#3489] Refactor Objects API to use new clients pattern
Instead of using zds-client/zgw-consumers clients, this is now building
directly on top of requests.Session.

The client factory in zgw_consumers typing is fixed to use a generic
for the client factory, signifying a more specific subclass is being
returned.

The DRC client integrations are also cleaned up, with the types now
using a protocol to satisfy the legacy zgw registration plugin clients,
while being compatible with the new approach. The goal is to move
the API-specific bits to the specific clients, allowing us to
normalize this when ZGW registration is refactored.

This will then in turn clean up a lot of the openforms.contrib.zgw
service module too.