:green_heart: Update tests
WebTest alone is no longer sufficient since analytics are enabled/disabled
on the fly using Javascript. The upside is that our cookie banner is now
more accessible, reference used is:
https://medium.com/@web-accessibility-education/dos-and-donts-of-accessible-cookie-banners-2c1a3102df2f

The analytics tools rendering tests no longer need an explicit accept to
have the script nodes in the DOM - they are inside the <template>
elements which are only activated after accepting, but this makes it
slightly easier to assert that the snippets are at least emitted.