:recycle: [open-formulieren/open-forms#3178] Replace layout wrapper with utrecht-page-content
The NL DS component provides in paddings around the content, so
our layout component in the backend is obsolete.

There is still a layout component in the SDK for the time being,
and the paddings of that one need tweaking so that the A11Y toolbar
and the form content properly align again.