:bug: [#3577] Relax CSP form-action directive which was too strict
While only allowing 'self' makes sense if the forms are only hosted on
our own backend, this breaks forms that are embedded on third party
domains. These domains need to be allow-listed in the CORS (and CSRF
trustlist) policy anyway, so we can use the explicit allow list as
a convenience 'automatic' configuration.

If this is not sufficient, at the infrastructure level the environment
variable 'CSP_EXTRA_FORM_ACTION' can be provided to add extra sources,
and as a last resort dynamic configuration through the admin interface
is possible.