:green_heart: Fix existing tests
The amount of get_metadata calls is changed because we directly check the
metadata file now instead of fetching a potentially unconfigured
source.

Additionally, .env overrides for the CSP_FORM_ACTION setting can cause
weird local test failures, so we pin that setting the tests that make
assertions against the response header.