:whale: [#3593] Install NodeJS v16 in production docker image
The NodeJS runtime is required to be able to run style-dictionary scripts, as there is no
Python implementation (yet).

Special care is taken to install NodeJS v16, since Debian bookworm ships NodeJS v18 in
its own repositories. Therefore, we pin this in the apt package manager. This version
needs to be kept in sync with our .nvmrc so that we use the same Node version for
all our (frontend) build tooling.

The alternative is building this in a separate project with its own container image,
which impacts the ease of deployment significantly. Style-dictionary itself also
relies on files being present on the filesystem, so adding an API layer to only
create and read files seems a bit convoluted. When style-dictionary v4 releases,
this may change.