:lock: Add CSP object-src directive to settings
This configuration was missing (but as a workaround it can be added
through the admin interface). Now we include it, which restricts
some functionality (which was not intended in the first place).

The envvar lookup/fallback allows an escape hatch for those who are
making use of object and embed tags.