:recycle: [#3620] Provide file component context
The form builder requires additional context for available file types,
confidentiality levels and callbacks to retrieve the available
document types for registration. All the configuration is readily
available in the DOM already, and the custom options for the formio
component implementation is broken out into a function shared
between the new and old form builder. It's also been refactored
to use async/await as much as possible and type information is
added via JSDoc annotations for improved readability.

Note that the Django templates are updated to ensure the
configuration script tags are emitted *before* the Javascript
modules are loaded/executed, so that the configuration is picked
up and we don't need an additional callable indirection to query
the DOM at runtime - which is best avoided inside React components
anyway.