:card_file_box: [#3403] Add support for form-specific theme configuration
This sets up  the database to point to a particular theme for a given
form.

Fixes an admin booboo too, 'list_select' is not a supported attribute
and using prefetch_related is likely more performant anyway since we
only expect small amounts of categories and themes, so a select_related
would worsen performance by throwing too much data over the wire.