🐛[#3672] Fixes Django template context in service fetch
Bluntly casting everything to str gives "lists" like `"[1, 2, 3]"`.
and `{{mylist.0}}` results in `"["`, not `1`.

Unsure if `datetime.*` should be cast to string or kept in tact,
because I don't know exactly what's meant with:

  # Explicitly cast values to strings to avoid localization

which was added in the switch from simple `{ }` interpolation  to the
Django template double bracket `{{ }}`

Maybe it was just to avoid 1.1 -> "1,1"?
I guess neither this `str(datetime.datetime.now())` nor the default that
Django template rendering gives, are the iso form an api probably wants.
But strings don't play well with the `|date` filter in the templates.