:recycle: [#3005] Refactor on submission completion task orchestration
Now when a submission is completed, when a cosign is done, when a payment is completed or a retry flow is triggered (either because the registration failed or a payment status update failed), the same task chain is called. This is to try and keep a maintainable overview of where tasks come from.