:recycle: [#3694] Centralize common component wrapping when mounting roots
Most of the time we're passing the same/similar context to the component
we're actually rendering, so it makes sense to make this a common wrapper
component.

The form-creation-form stuff doesn't play nice with strict mode yet, there
are unintended side-effects being performed in places (formio.js for sure)
that trip immer. So, for the time being, strict mode can be disabled.