:green_heart: [#3049] Fix broken WebTest tests
The tests no longer pass due to how the form widget (Django) and
JS interact, as WebTest does not execute any Javascript. No
initial (valid and JSON-decodable) value is rendered, which causes
Django to crash on the empty inputs when trying to parse it as
JSON.

As a workaround, we can just grab the instance values and manually
serialize them to simulate the behaviour of django-jsonform.