:card_file_box: [#3733] Increase scope of data migration, fix broken file components
Ran a detection script on some of our environments and it seems that
there is no structural problem with these file component default values.

It is causing Javascript crashes however, so corrupt data needs to be fixed.

None of our form builders allow setting a default value for the file upload,
and both an empty array or 'null' as default value *should* work. The
behaviour of the form builder is applied here, which is to set the default
value to 'null' in all cases (multiple true/false).