:bug: [#3749] Replace useEffect with useAsync
In react 18, there has been this change (https://github.com/facebook/react/blob/main/CHANGELOG.md#breaking-changes):

```
Consistent useEffect timing: React now always synchronously flushes effect functions if the update was triggered during a discrete user input event such as a click or a keydown event. Previously, the behavior wasn't always predictable or consistent.
```

normally the callback function of the useEffect should return the cleanup function. However, in the LogicInputDescription we had an async callback (which returns a promise). Then when clicking on the logic rule description this would trigger the error `Uncaught TypeError: destroy is not a function` . I think this happens because it tries to use the promise as the cleanup function.