:bug: Fix tests and handling of existing component definitions
The 'component' argument in editComponent (Formio webformbuilder)
does not contain the validate key, we need to ensure that the
defaults are copied in from the default schema.

I'm unsure how this relates to the 'original' argument and how
exactly it's being used, but it doesn't seem that we're losing
any other data at the moment (tested with nested translations).

The test data in the e2e tests also needs to be updated to
properly reflect the stricter schema constraints we now apply.