:thread: [#3858] Fix concurrent session updates with multiple uploads
Multiple uploads fire in parallel by the browser, which apparently
causes the views/middleware to start processing the session data.

This means that when an upload completes and it writes the
upload UUID to the session, that each upload thinks there
were no previous uploads, so it only writes itself to the
session. The last completed upload is then the only one saved in
the session data, and available for permission checks.

Due to earlier uploads not being present in the session data,
deleting those results in HTTP 403 errors, as we use the
session in this case for access control.

Starting multiple submissions (for different forms) would
have had the same problem if they happen in quick succession,
so we may have fixed another unreported problem.

Note that updating the session here is very tricky code -
however using some kind of "signed URL" is no proper
solution either. We already use near-impossible to guess
UUIDs in the URLs, but the risk remains the same: if you
have the URL, you can read/delete potentially sensitive
data that should only be available to the uploader. So,
we *need* to maintain server-side state to guarantee
this.

As for the fix itself, I've opted to not solve this in
a generic way that would allow concurrency on session
data in general, the fact that Django doesn't handle this
out of the box seems like good warning about the complexity
involved in solving such a problem. In particular, tracking
which changes are deliberate and which changes are stale
does not seem feasible. Instead, a very localized solution
is applied in problem areas known to suffer from concurrency
issues - for which the bulk upload of files in the browser
is probably the most common.

The locking parameters have been guesstimated, taking into
account current browser behaviour in amount of parallel
requests they actually execute (typically 6). Local profiling
shows that the actual lock acquiring is very fast. This
makes redis a hard-dependency in the codebase, but we didn't
really support alternative cache backends anyway.