:alien: [#3049] Correctly derive which theme to apply in tinymce
Due to auto/explicit modes now being a thing, the detection purely through
a media query is not sufficient. Instead, if an explicit theme is set,
that maps to the skin/content_css properties.

If the theme mode is set to 'auto', then the media query applies to
figure out whether dark or light theme needs to be applied.

The code is refactored to remove duplication and instead have a more
declarative approach for possible future themes.