:recycle: [#3049] Redo the tooltip styling from scratch
The Django markup changed substantially so the existing CSS was no
longer usable. However, using the :has pseudo selector we can
re-instate the same appearance with some clever CSS grid/flexbox
usage.

Colors for the tooltip are now specified as CSS variables too, so
this patch should be easily backported to default-project. The
relevant styles are put in a separate SASS module to easier keep
track of which aspect of the django admin they touch.