:bug: [#3924] Update payment status also if 'wait_for_payment_to_register=True'
The problem was:
If the registration happens after payment, the status of the payment was never changed from 'completed' to 'registered'. The 'finalise_completion' task filters for payments with 'registered' status to decide whether the retry flag should be turned on. This resulted in the retry flag being True even though the payment has been registered.