:sparkles: [#3977] Implement the (nested) editgrid component to serializer field
The editgrid represents a nested serializer which is repeated (array type),
so the serializer field is a list field wrapping the recursively build
nested serializer from the components making up the edit grid.

Validations like max number of items, required... are implemented
too. Because this is a nested serializer, each item in the array
is individually validated too.