:safety_vest: [#3977] Run formio component validation during submission complete
The final (dynamic) formio form/components configuration is
translated into a serializer instance, which is used to perform
a final input validation step of the submitted user input.

This ensures that invalid data is not possibly submitted
to any registration backend, while allowing features like
form pausing to still work.