:adhesive_bandage: [#2827] Remove seconds/microsecond resolution from 'now' static variable
This resolution is typically not important, as the frontend UI widgets
also already truncate the seconds (bringing the time/datetime resolution
to minutes).

Ensuring that multiple logic check calls within the same minute yield
the same value breaks the logic check/check schedule infinite loop and
resolves the reported problem. A better fix would be to prevent these
loops in the SDK in the first place, but for that we need the new
renderer with a smarter way to do the logic checks.