:bug: [DH#673] Fix incorrect 'empty' values being included.
While the original bug report expects/states that the hidden fields should
not be included at all in the objects API data, this is not actually a
bug but a deliberate action - it was patched in #3890. It was never
intentional to *not* send (the values) of hidden fields.

This does however mean that the defined object type JSON schema's
need to be updated to account for possible empty values, where
the data type of empty is determined by the component type of
the field (e.g. number and currency are 'null', but textfield is
always going to be string).

What *was* broken however, is that missing values were always
sent as `null\' rather than the empty value that's component
specific.

There is an additional report of an entire fieldset being
hidden resulting in the values not being sent, so that will
require patching too.

Backport-of: #4123