:bug: [#4126] Rework editgrid serializer field setup
We cannot simply use ListField, as that applies the same
child serializer for every item inside the array of values,
but the field configuration of the child actually changes
depending on the values passed in (some fields may be
hidden, and that disables their validations entirely).

Note the disclaimer about rest_framework.fields.Field.__repr__,
this method behaves unexpectedly because it doesn't report
the current values of args/kwargs, but the values at the
time of instance creation.