:green_heart: reset cache after flaky test detection
The cache is in a 'broken' state, which causes any additional
retries to fail because of stale data being present. For
proper robustness, the cache must be cleared between retries.

It *could* be that there's a cache issue causing the flakiness
in the first place, so that's why it's not being cleared
at the beginning of the test - making sure we can continue
investigating the source of flakiness.