:art: [#4236] Remove unneeded settings patching
By specifying a registration backend, we don't need to
check the GlobalConfiguration solo model (via the
factories) and thus don't risk accessing any cached
data that could break test isolation.

It's a bit weird that this factory helper does DB
and cache IO to select an enabled plugin, we
should probably clean that up in a later stage.