:recycle: [#4246] Refactor digig-eherkenning-oidc-generics auth request flow
Simplified the entire authentication request flow where the user
gets redirect to the relevant identity provider.

There is now a single 'view' implementation that takes a config
class/model to use which can be used to directly obtain the
redirect target instead of having to go through multiple
redirects on our own URLs.

The view takes care of input sanitation and managing the
authentication state.

This substantially cleans up the inheritance/mixin chains for
the OIDC flows and makes the code easier to follow.