:sparkles: [#4269] Process input expressions from the DMN engine
Input expressions can be complex with the actual requested variable buried
deep inside. We leverage the 'feelin' library to parse FEEL expressions
so that we can extract the actual (likely) expected variables and
combine this information with our retrieved input clauses.

This result is used to build a list of DMN variables for input
mapping - this list is ordered to put labeled literal variables
at the top, followed by the extracted unlabeled variables, and
finally the legal identifiers that are labeled but likely
complex/edge cases.

This implements the DMN 1.3 FEEL identifier specification to
select candidates.