:card_file_box: [#3807] Add model field for co-sign request template
Added the model field and necessary data migration to
expose the co-sign request email content as a template
to the administrators.

The data migration sets up the default templates as they
used to be in their hardcoded form, taking into account
the option to include/exclude direct links to forms.