:arrow_up: Upgrade requests
Addresses CVE-2024-35195. This vulnerability should not be a problem
for us either way since we don't make calls with verify=False, but
upgrading should not be a problem either.

Opting into verify=False is only possible through some envvars that
disable verification entirely and leads to an insecure configuration
either way, which we don't recommend for production purposes.