:bug: [#4278] Fix form values not being reset on dmn version change
Now when the decision definition version changes and this results in
different input/output parameters, the internal form state is updated
to properly reset stale references.

Before this patch, the dropdown would render as if nothing was
selected because the active value is/was not present in the options,
but under the hood the form data was still holding on to this value
and there would not be a way to reset this (or see that it's broken).

We now check for stale references and reset these if we encounter
them.