:card_file_box: [#4246] Expand AuthInfo model to capture mandate context
* Added fields for the (optional) legal subject (authorizee)
* Added fields for the (optional) acting subject (authorizee)
* Added check constraints to enforce data integrity
* Organized the admin in logical groups

The authentication context data model identifies two parties:
the representee and the authorizee. There is always an
authorizee - it *can* be the same actor as the representee
if no mandate is in involved.

An authorizee has two aspects: legal subject and acting
subject. There is always a legal subject. If no acting
subject is provided, it is inferred from the legal subject.

So, a simple DigiD login in this model is represented by
a legal subject authorizee of type BSN. We still keep
storing this data in the attribute + value fields, but
in the event of a mandate, we will store the additional
information for the legal/acting subject.