:hammer: [#4341] Fix api mock base URLs
The deployed storybook on a subpath needed some workarounds for mocks
in the past, because the API endpoints call absolute paths rather than
fully qualified URLs. This is because github pages hosts on a subpath,
however Chromatic hosts on a sub-domain for each build so the same
solution does not work.

Instead, we can now build storybook with a dedicated 'mock domain'
for API endpoints, with a config that is picked up by the fetch
wrapper and storybook mocks. If not specified, it defaults to the
current behaviour.