:lipstick: Build storybook in light mode
We don't build the admin overrides CSS, which sets the design token
values for dark mode, so the 'auto' mode in dark themed
browsers looks horrible in the deployed version.

Building in light mode at least makes this consistent, perhaps
at some point in the future this can be made dynamic for
local development.