:ambulance: [#4362] Prevent crash due to translatable defaultValue
Textfield and textarea have the multiple: true option, which turns the defaultValue
into an array of strings rather than a single string, and that
doesn't play well with the translations machinery or missing
translation detection.

Open Forms 2.7.0 removes support for translations for this
property, and older versions get a hotfix to prevent the
crash from happening during detection.