:bug: [#4338] Fix SOAP 1.2 namespace replacement
Not the scattering of namespaces is an issue, but our
mapping of namespace replacements. We replace the
namespaces to get easier dict lookups (without ns
URLs prefixed to keys), but we only covered the
SOAP 1.1 Envelope namespace.

The XML message that causes issues uses the XML
namespace for SOAP 1.2 Envelope, and since that
namespace was not being replaced/removed, the
root 'Envelope' key in the dict does not exist
(it only exists with the namespace prefix URL),
causing all nested lookups to fail and no values
to be returend.

This patch adds the SOAP 1.2 namespace, which
ensures this gets replaced and now the root key
can be found, including all nested keys and
data.