:whale: [#4246] Redo/setup docker-compose Keycloak
The previous configuration was geared towards running Keycloak locally
to test out full integration (manually, with runserver). This was not
really used by anyone, so the configuration got outdated.

Instead, the setup is now done so that the config is spun up every time
we re-record VCR cassettes. The Client ID and Secret are included in
the VCR.py tests and we record the full network traffic for OIDC-based
tests during tests to replay them on CI.

The README documents how to dump the 'fixture' again after making
changes in Keycloak's UI for when requirements change.