:arrow_up: [#4246] Upgrade mozilla-django-oidc-db and digid-eherkenning
Added the digid-eherkenning[oidc] optional dependency group, which
includes the configuration models for DigiD/eHerkenning. This code was
extract from src/digid_eherkenning_oidc_generics and is (almost) a
drop-in replacement for Open Forms.

Also upgraded the mozilla-django-oidc-db package to make use of the
configuration/settings rework done there, so that we can cut down on
amount of code needed in our views/plugins.