:recycle: [#4246] Rework the organization-oidc authentication plugin
The plugin has been reworked to make use of the new mechanics for
configuration customization in mozilla-django-oidc-db. The tests are
restructured to make use of VCR.py instead of manual mocks, giving
better confidence that things are actually working.

This should make the code and tests a lot more maintainable.