:card_file_box: [#4246] Define custom configuration (proxy) models
The digid_eherkenning.oidc package defines the relevant models and
configuration to handle DigiD/eHerkenning via OIDC - the code in there
was moved over from Open Forms and cleaned up. Additional configuration
is added there to be able to capture the full authentication context -
this follows the specification written for the authentication context
data model.

We subclass those config models (as proxy models) to change the Python
behaviour without touching the DB tables:

* set up mechanism to use the legacy callback URLs, since these are
  registered with the identity provider and if we were to just change
  our URLs, that will cause outages. Users can opt into the new
  behaviour - a better construction via feature flags will be added
  later
* Add properties for the 'mandate claims' to extract more information
  about 'machtigen' from the configured claim fields. This property
  will be consumed in our custom authentication backend.
* Register a custom admin form so that the invididual configurations
  can only be disabled if no (Open Forms) forms make use of the
  respective authentication plugins.